requirements.txt
# --- Web Framework -----------------------------------------------------------
flask>=3.0
python-dotenv>=1.0
werkzeug>=3.0
requests>=2.31

# --- Database ----------------------------------------------------------------
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
psycopg2-binary>=2.9  # PostgreSQL adapter for Neon
Flask-Migrate>=4.0
Flask-Login>=0.6

# --- Deployment --------------------------------------------------------------
gunicorn>=21.2

# --- ML / Computer Vision ----------------------------------------------------
# mediapipe>=0.10.30 required for Python 3.14+ compatibility on Render
mediapipe>=0.10.30
opencv-python>=4.9

# --- Data Science ------------------------------------------------------------
numpy>=1.26
pandas>=2.2
scipy>=1.12
scikit-learn>=1.4
pillow>=10.2

# --- Testing -----------------------------------------------------------------
pytest>=8.0
